Changes priority from int to double

+++ AI/p_queue.h
@@ -15,11 +15,11 @@ class p_queue
     public:
         p_queue(int width, int height);
         p_queue& operator=(p_queue& a);
-        bool lessPriority(std::pair<Point, int> &p1, std::pair<Point, int> &p2);
+        bool lessPriority(std::pair<Point, double> &p1, std::pair<Point, double> &p2);
         void push_up_heap(int index);
         void push_down_heap(int index);
         void swap_nodes(int a, int b);
-        void insert(Point& x, int p);
+        void insert(Point& x, double p);
         Point& pop();
         void ndelete(Point& P);
         Point& top();
@@ -27,12 +27,12 @@ class p_queue
         bool contains(Point& key);
         bool compPoints(Point& a, Point& b);
         int getSize();
-        std::vector<std::pair<Point,int>>* getContainer();
+        std::vector<std::pair<Point,double>>* getContainer();
 
       
 
     private:
-        std::vector<std::pair<Point, int>>* container;
+        std::vector<std::pair<Point, double>>* container;
         std::vector<std::vector<int>> indirection;
 
         std::pair<Point, int> getLeftNode(int currindex);
